refactor(admin): move query-string parsing out of the route registry's reach
Reading a query string had to be imported from `lib/routing`, which resolves a
path against the page registry and therefore imports every page in the admin.
Anything a PAGE uses could not reach those helpers without closing a cycle back
onto itself — `useEditorLocale` → `useSearchParams` → `lib/routing` →
`pages/registry` → the entry page → `useEditorLocale`.

The fallow gate caught it: two circular dependencies introduced, verdict fail.

Parsing a query string has nothing to do with knowing what routes exist, so it
moves to `lib/search-params`, which imports nothing. `lib/routing` re-exports
the three names so existing callers are unchanged.

That also closes a cycle that was already there before this branch, through
`EntryList`. Fifteen remain, all through `useRouter` and none of them this
change's.

The locale query key moves to `constants/search-params` for the same reason from
the other direction: the entry list writes the key and the editor reads it, so a
constant defined beside either one drags that side's imports into the other.
