+++ packages/admin/src/components/features/entries/EntryList/EntryList.tsx
@@ -19,13 +19,13 @@ import { Breadcrumbs } from "@admin/components/shared";
 import { PluginSlot } from "@admin/components/shared/plugin-slot";
 import { useListColumns } from "@admin/components/ui/table/list-view";
 import { ROUTES, buildRoute } from "@admin/constants/routes";
+import { LOCALE_PARAM } from "@admin/constants/search-params";
 import {
   useEntries,
   useBulkDeleteEntries,
   useBulkUpdateEntries,
 } from "@admin/hooks/queries";
 import { useCollection } from "@admin/hooks/queries/useCollections";
-import { LOCALE_PARAM } from "@admin/hooks/useEditorLocale";
 import { useEntryListShortcuts } from "@admin/hooks/useKeyboardShortcuts";
 import { useLocalization } from "@admin/hooks/useLocalization";
 import { usePluginAutoRegistration } from "@admin/hooks/usePluginAutoRegistration";
@@ -33,7 +33,7 @@ import { useSearchParams } from "@admin/hooks/useSearchParams";
 import type { ListResponse } from "@admin/lib/api/response-types";
 import { navigateTo } from "@admin/lib/navigation";
 import type { InjectionPointProps } from "@admin/lib/plugins/component-registry";
-import { getSearchParam } from "@admin/lib/routing";
+import { getSearchParam } from "@admin/lib/search-params";
 import type { Entry } from "@admin/types/collection";
 import type { ApiCollection } from "@admin/types/entities";
